docker/unet/requirements.txt
pandas
numpy
scipy
pillow
keras
scikit-learn
scikit-image
opencv-python
openslide-python